Improved findSkillNode algorithm.
Algorithm no longer searches through the tree twice. Instead,
it makes a tradeoff of declaring a local variable inside each
instance. The number of local variables is determined by the depth
of the search. In any case, I view this as better than having no
local variables but having to search the tree twice. The code
also looks better, incidentally.